requirements.txt
ks-csc-ai-models
json5~=0.12.1
pandas~=2.0.3
infra-framework
setuptools~=75.1.0
flask~=3.0.3
grpcio~=1.59.2
protobuf~=3.20.1
pydantic~=2.9.2
diskcache~=5.6.3
requests~=2.32.4
tiktoken~=0.7.0
jsonschema~=4.25.0
dashscope~=1.24.1
langchain-core~=0.2.43
pytest~=8.3.5
dataclasses~=0.6
json_repair~=0.30.0
numpy~=1.24.4
openpyxl~=3.1.5
mammoth~=1.11.0